新規登録表示

+++ src/main/java/com/coltware/spring/repository/ProductRepository.java
@@ -8,6 +8,7 @@ import com.coltware.spring.model.Product;
 
 @Repository
 
-public interface ProductRepository extends JpaRepository<Product, Integer>, JpaSpecificationExecutor<Product> {
+public interface ProductRepository extends JpaRepository<Product, Long>, JpaSpecificationExecutor<Product> {
+
 
 }
